fix promotion

+++ src/main/java/com/app/TicketBookingMovie/dtos/PromotionDetailDto.java
@@ -7,7 +7,7 @@ public class PromotionDetailDto {
     private Long id;
     private String typeDiscount;
     private double discountValue;
-    private double maxValue;
+    private int maxValue;
     private double minBillValue;
     private Long foodId;
     private Long promotionLineId;
